moved helper functions into test_utils
